PlayStationAPI/requirements.txt
Flask
pymongo
python-dotenv
Flask-Cors
gunicorn